# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Iraqi School Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List all students
      parameters:
        - name: classId
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: createStudent
      tags: [students]
      summary: Create a student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Not found
    patch:
      operationId: updateStudent
      tags: [students]
      summary: Update a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Delete a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /teachers:
    get:
      operationId: listTeachers
      tags: [teachers]
      summary: List all teachers
      parameters:
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of teachers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Teacher"
    post:
      operationId: createTeacher
      tags: [teachers]
      summary: Create a teacher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeacherInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
  /teachers/{id}:
    get:
      operationId: getTeacher
      tags: [teachers]
      summary: Get a teacher
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Teacher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
        "404":
          description: Not found
    patch:
      operationId: updateTeacher
      tags: [teachers]
      summary: Update a teacher
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeacherUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
    delete:
      operationId: deleteTeacher
      tags: [teachers]
      summary: Delete a teacher
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /subjects:
    get:
      operationId: listSubjects
      tags: [subjects]
      summary: List all subjects
      responses:
        "200":
          description: List of subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subject"
    post:
      operationId: createSubject
      tags: [subjects]
      summary: Create a subject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
  /subjects/{id}:
    patch:
      operationId: updateSubject
      tags: [subjects]
      summary: Update a subject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubjectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
    delete:
      operationId: deleteSubject
      tags: [subjects]
      summary: Delete a subject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Student:
      type: object
      required: [id, fullName, studentCode, classId, className, gender, dateOfBirth, enrollmentDate, status]
      properties:
        id: { type: integer }
        fullName: { type: string }
        studentCode: { type: string }
        classId: { type: integer }
        className: { type: string }
        gender: { type: string }
        dateOfBirth: { type: string }
        enrollmentDate: { type: string }
        phone: { type: ["string", "null"] }
        parentName: { type: ["string", "null"] }
        parentPhone: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        status: { type: string }
    StudentInput:
      type: object
      required: [fullName, classId, gender, dateOfBirth]
      properties:
        fullName: { type: string }
        classId: { type: integer }
        gender: { type: string }
        dateOfBirth: { type: string }
        phone: { type: string }
        parentName: { type: string }
        parentPhone: { type: string }
        address: { type: string }
        status: { type: string }
    StudentUpdate:
      type: object
      properties:
        fullName: { type: string }
        classId: { type: integer }
        gender: { type: string }
        dateOfBirth: { type: string }
        phone: { type: string }
        parentName: { type: string }
        parentPhone: { type: string }
        address: { type: string }
        status: { type: string }
    Teacher:
      type: object
      required: [id, fullName, teacherCode, specialization, phone, status]
      properties:
        id: { type: integer }
        fullName: { type: string }
        teacherCode: { type: string }
        specialization: { type: string }
        phone: { type: string }
        email: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        hireDate: { type: string }
        status: { type: string }
    TeacherInput:
      type: object
      required: [fullName, specialization, phone]
      properties:
        fullName: { type: string }
        specialization: { type: string }
        phone: { type: string }
        email: { type: string }
        address: { type: string }
        hireDate: { type: string }
        status: { type: string }
    TeacherUpdate:
      type: object
      properties:
        fullName: { type: string }
        specialization: { type: string }
        phone: { type: string }
        email: { type: string }
        address: { type: string }
        hireDate: { type: string }
        status: { type: string }
...
    Subject:
      type: object
      required: [id, name, teacherId, teacherName, classId, className, weeklyHours]
      properties:
        id: { type: integer }
        name: { type: string }
        teacherId: { type: integer }
        teacherName: { type: string }
        classId: { type: integer }
        className: { type: string }
        weeklyHours: { type: integer }
        description: { type: ["string", "null"] }
    SubjectInput:
      type: object
      required: [name, teacherId, classId, weeklyHours]
      properties:
        name: { type: string }
        teacherId: { type: integer }
        classId: { type: integer }
        weeklyHours: { type: integer }
        description: { type: string }
    SubjectUpdate:
      type: object
      properties:
        name: { type: string }
        teacherId: { type: integer }
        classId: { type: integer }
        weeklyHours: { type: integer }
        description: { type: string }
...