# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Iraqi School Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface Student {
  id: number;
  fullName: string;
  studentCode: string;
  classId: number;
  className: string;
  gender: string;
  dateOfBirth: string;
  enrollmentDate: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  parentName?: string | null;
  /** @nullable */
  parentPhone?: string | null;
  /** @nullable */
  address?: string | null;
  status: string;
}