# lib/api-zod/src/generated/types/studentUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Iraqi School Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface StudentUpdate {
  fullName?: string;
  classId?: number;
  gender?: string;
  dateOfBirth?: string;
  phone?: string;
  parentName?: string;
  parentPhone?: string;
  address?: string;
  status?: string;
}