# lib/api-zod/src/generated/types/subjectUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Iraqi School Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface SubjectUpdate {
  name?: string;
  teacherId?: number;
  classId?: number;
  weeklyHours?: number;
  description?: string;
}