# lib/api-zod/src/generated/types/teacherInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Iraqi School Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface TeacherInput {
  fullName: string;
  specialization: string;
  phone: string;
  email?: string;
  address?: string;
  hireDate?: string;
  status?: string;
}